Final Traverse the Earth
Adjusted the Form for some responsiveness on the Contact us page in the HTML and the JS.
It could use some tweaking but it generally works.  Updated more CSS styling. Updated queries for a bit more responsiveness but would appreciate some input on how to improve. Generally cleaned up alignment in the code and comments.  Any comments really would be appeciated.  Thanks!